[Refactor/#45] Orum ver 1.0 Refactori ng
