--- Orum/Orum/Models/Model.swift
@@ -1,8 +1,0 @@
-//
-//  Model.swift
-//  Orum
-//
-//  Created by 차차 on 10/16/23.
-//
-
-import Foundation
