--- Orum/Orum/Views/ContentView.swift
@@ -1,24 +1,0 @@
-//
-//  ContentView.swift
-//  Orum
-//
-//  Created by 차차 on 10/16/23.
-//
-
-import SwiftUI
-
-struct ContentView: View {
-    var body: some View {
-        VStack {
-            Image(systemName: "globe")
-                .imageScale(.large)
-                .foregroundStyle(.tint)
-            Text("Hello, world!")
-        }
-        .padding()
-    }
-}
-
-#Preview {
-    ContentView()
-}
